Move test from helpers module into tests directory
